requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
Django==3.1.4
django-admin-thumbnails==0.2.5
django-authtools==1.7.0
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.9.0
django-colorfield==0.4.2
django-constance==2.8.0
django-domainname==0.1.1
django-faicon @ git+https://github.com/dev-rathankumar/django-faicon-1.git@215de0c10cac7a4f9d56df741c5cb2e4f506da79
django-formset-js==0.5.0
django-formtools==2.3
django-google-translate==1.1
django-jquery-js==3.1.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-sekizai==2.0.0
django-smart-selects==1.5.9
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
filelock==3.0.12
idna==3.2
numpy==1.21.2
Pillow==8.3.2
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2020.4
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-file==1.5.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tldextract==3.1.2
urllib3==1.26.2
gunicorn==20.1.0
